Puppet support
This update brings puppet provisioning support. Point PUPPET_SERVER
to any puppetmaster and it activates the install of puppet and adds a
basic config with PUPPET_SERVER. If you also set PUPPET_CONF it will
download your custom config.

This update also revert a fix for a bug in grub that now seems to be
resolved.